# Provide a faster fallback for memrchr