There are patches in the pipeline for 16-bit pointer sizes and so on in the main rust tree, so that got me thinking about making sure it compiles no matter what the target pointer size is. Specialization is only for 32-bit and 64-bit platforms.

benchmark looks like this with the fallback used for memrchr, tested on 64-bit linux (fallback is normally not used in linux, but here it is).

```
test iterator          ... bench:      11,799 ns/iter (+/- 36) = 847 MB/s
test iterator_reversed ... bench:      11,781 ns/iter (+/- 29) = 848 MB/s
test libc_memchr       ... bench:         455 ns/iter (+/- 3) = 21978 MB/s
test libc_memrchr      ... bench:       2,236 ns/iter (+/- 12) = 4472 MB/s
```