@bluss 2015-08-19
Oh :laughing:  memrchr fallback is [faster](https://ci.appveyor.com/project/BurntSushi/rust-memchr/build/1.0.8/job/bddb7c3amyy7cdp3) than memchr itself on windows according to the appveyor tests..

I can of course provide the same algorithm in its forward direction, if wanted.