@BurntSushi 2015-08-19
@bluss What the..? Really? If you want to, sure, that benchmark seems pretty convincing...