@bluss 2015-08-19
Found an improvement.. new bench is

```
test libc_memrchr      ... bench:       1,974 ns/iter (+/- 26) = 5065 MB/s
```

just slightly better, but this one actually vectorized on my tested platform.