@bluss 2015-08-19
I'll come back to memchr another day. This is fine to merge if you want.

For memchr, I wonder what the OS X situation looks like?