@BurntSushi 2015-08-19
@bluss Thanks!

No clue about OS X. I have a Mac mini next to me that has been sitting in its box since December. Been meaning to get it set up to build/dist Rust programs...