@nicokoch 2015-08-19
Just a heads up: I tested this implementation on my own windows box and @bluss's implementation is clearly faster than windows' memchr:
```
running 4 tests
test iterator          ... bench:       4,767 ns/iter (+/- 3,443) = 2097 MB/s
test iterator_reversed ... bench:       4,630 ns/iter (+/- 1,386) = 2159 MB/s
test libc_memchr       ... bench:       4,062 ns/iter (+/- 1,345) = 2461 MB/s
test libc_memrchr      ... bench:         870 ns/iter (+/- 230) = 11494 MB/s

test result: ok. 0 passed; 0 failed; 0 ignored; 4 measured
```