@BurntSushi 2015-08-19
@bluss I knew you were good... but I didn't know you were *that* good!