@bluss 2015-08-20
wow! I want to thank llvm for the codegen :wink: